refactor(assignments-service): Add FileModule and FileService
