+++ services/apps/assignments/src/classroom/classroom.service.ts
@@ -4,14 +4,12 @@ import {Method} from 'axios';
 import {createReadStream} from 'fs';
 import {firstValueFrom} from 'rxjs';
 import {Stream} from 'stream';
-import {Entry as ZipEntry, Parse as unzip} from 'unzipper';
 import {AssignmentDocument} from '../assignment/assignment.schema';
-import {SearchService} from '../search/search.service';
 import {AuthorInfo, Solution} from '../solution/solution.schema';
 import {SolutionService} from '../solution/solution.service';
 import {generateToken} from '../utils';
-import {MAX_FILE_SIZE, TEXT_EXTENSIONS} from "../search/search.constants";
 import {ImportSolution} from "./classroom.dto";
+import {FileService} from "../file/file.service";
 
 interface RepositoryInfo {
   name: string;
@@ -30,7 +28,7 @@ interface SearchResult {
 export class ClassroomService {
   constructor(
     private solutionService: SolutionService,
-    private searchService: SearchService,
+    private fileService: FileService,
     private http: HttpService,
   ) {
   }
@@ -57,7 +55,7 @@ export class ClassroomService {
       await Promise.all(files.map(async (file, index) => {
         const stream = createReadStream(file.path);
         const solution = solutions.upsertedIds[index];
-        return this.importZipFiles(stream, assignment.id, solution);
+        return this.fileService.importZipEntries(stream, assignment.id, solution);
       }));
     }
 
@@ -97,33 +95,6 @@ export class ClassroomService {
     } else {
       return ['name', filename.slice(0, -4)];
     }
-  }
-
-  private async importZipFiles(stream: Stream, assignment: string, solution: string, commit?: string) {
-    await stream.pipe(unzip()).on('entry', (entry: ZipEntry) => {
-      // Using vars.uncompressedSize because entry.extra.* and entry.size are unavailable before parsing for some reason
-      if (entry.type !== 'File' || (entry.vars as any).uncompressedSize > MAX_FILE_SIZE) {
-        entry.autodrain();
-        return;
-      }
-      const extension = entry.path.substring(entry.path.lastIndexOf('.') + 1);
-      if (!TEXT_EXTENSIONS.has(extension)) {
-        entry.autodrain();
-        return;
-      }
-      entry.buffer().then(buffer => {
-        if (buffer.length > MAX_FILE_SIZE) {
-          return;
-        }
-        const strippedPath = commit ? this.stripCommit(entry.path, commit) : entry.path;
-        this.searchService.addFile(assignment, solution, strippedPath, buffer.toString('utf8'));
-      });
-    }).promise();
-  }
-
-  private stripCommit(filename: string, commit: string) {
-    const index = filename.indexOf(commit);
-    return index >= 0 ? filename.substring(index + commit.length + 1) : filename;
   }
 
   async countSolutions(assignment: AssignmentDocument): Promise<number> {
@@ -178,7 +149,7 @@ export class ClassroomService {
       const upsertedId = solutions.upsertedIds[i];
       if (commit && upsertedId) {
         const zip = await this.getRepoZip(assignment, this.getGithubName(repo, assignment), commit);
-        return zip ? this.importZipFiles(zip, assignment._id, upsertedId, commit) : [];
+        return zip ? this.fileService.importZipEntries(zip, assignment._id, upsertedId, commit) : [];
       } else {
         return [];
       }
@@ -202,7 +173,7 @@ export class ClassroomService {
     return repositories
       .map(repo => this.createImportSolution(assignment, repo, undefined))
       .sort((a, b) => a.author.github!.localeCompare(b.author.github!))
-    ;
+      ;
   }
 
   private async getRepositories(assignment: AssignmentDocument): Promise<RepositoryInfo[]> {
