+++ services/apps/assignments/src/file/file.module.ts
@@ -1,0 +1,11 @@
+import {Module} from '@nestjs/common';
+import {SearchModule} from "../search/search.module";
+import {FileService} from "./file.service";
+
+@Module({
+  imports: [SearchModule],
+  providers: [FileService],
+  exports: [FileService],
+})
+export class FileModule {
+}
